Rename module: host_service -> hosts
