+++ bcs-app/backend/container_service/infras/hosts/constants.py
